requirements.txt
django-crispy-forms
crispy-bootstrap5
pandas
numpy
scikit-learn
pillow
django
djangorestframework
django-debug-toolbar
django-cors-headers
djangorestframework-simplejwt